server_pi_image_handling/server/Analysis/requirements.txt
imageio
PILLOW
pygifsicle
matplotlib